+++ app/src/main/java/com/github/geekarist/henripotier/Offer.java
@@ -4,40 +4,40 @@ import java.util.List;
 
 public class Offer {
     final Type type;
-    final int value;
+    final double value;
     final int sliceValue;
 
-    public Offer(Type type, int value) {
+    public Offer(Type type, double value) {
         this.value = value;
         this.type = type;
         this.sliceValue = 0;
     }
 
-    public Offer(String type, int value) {
+    public Offer(String type, double value) {
         this(Type.valueOf(type), value);
     }
 
-    public Offer(Type type, int sliceValue, int value) {
+    public Offer(Type type, int sliceValue, double value) {
         this.type = type;
         this.sliceValue = sliceValue;
         this.value = value;
     }
 
-    public int apply(List<Book> books) {
+    public double apply(List<Book> books) {
         if (type == Type.minus) {
             return value;
         } else if (type == Type.percentage) {
-            int sum = totalPrice(books);
+            double sum = totalPrice(books);
             return sum * value / 100;
         } else if (type == Type.slice) {
-            int sum = totalPrice(books);
-            return (sum / sliceValue) * value;
+            double sum = totalPrice(books);
+            return ((int) (sum / sliceValue)) * value;
         }
         return 0;
     }
 
-    private int totalPrice(List<Book> books) {
-        int sum = 0;
+    private double totalPrice(List<Book> books) {
+        double sum = 0;
         for (Book b : books) {
             sum += b.price;
         }
